#5 refactor: hub - DDD 구조로 변경
