+++ hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/application/dto/CreateHubResponse.java
@@ -1,9 +1,8 @@
-package com.fn.eureka.client.hubservice.hub.dto;
+package com.fn.eureka.client.hubservice.hub.application.dto;
 
 import java.util.UUID;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 
 @Getter
