--- hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/dto/ReadHubResponse.java
@@ -1,4 +1,0 @@
-package com.fn.eureka.client.hubservice.hub.dto;
-
-public class ReadHubResponse {
-}
