+++ hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/presentation/HubController.java
@@ -1,4 +1,4 @@
-package com.fn.eureka.client.hubservice.hub.controller;
+package com.fn.eureka.client.hubservice.hub.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.fn.eureka.client.hubservice.hub.dto.CreateHubRequest;
-import com.fn.eureka.client.hubservice.hub.dto.CreateHubResponse;
-import com.fn.eureka.client.hubservice.hub.service.HubService;
+import com.fn.eureka.client.hubservice.hub.application.HubService;
+import com.fn.eureka.client.hubservice.hub.application.dto.CreateHubRequest;
+import com.fn.eureka.client.hubservice.hub.application.dto.CreateHubResponse;
 
 import lombok.RequiredArgsConstructor;
