--- hub-service/src/main/java/com/fn/eureka/client/hubservice/hub/repository/HubRepository.java
@@ -1,10 +1,0 @@
-package com.fn.eureka.client.hubservice.hub.repository;
-
-import java.util.UUID;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.fn.eureka.client.hubservice.hub.domain.Hub;
-
-public interface HubRepository extends JpaRepository<Hub, UUID> {
-}
